backend/requirements.txt
fastapi
uvicorn[standard]

sqlalchemy
psycopg2-binary

python-dotenv
pydantic

passlib[bcrypt]
bcrypt==3.2.0

python-jose[cryptography]

python-multipart

pydantic[email]

